# Improve MapSet comparison performance