--- a/lib/elixir/lib/map_set.ex
+++ b/lib/elixir/lib/map_set.ex
@@ -158,7 +158,8 @@ defmodule MapSet do
       when map_size(map1) < map_size(map2) * 2 do
     map =
       map1
-      |> Map.keys()
+      |> :maps.iterator()
+      |> :maps.next()
       |> filter_not_in(map2, [])
 
     %MapSet{map: map}
@@ -171,13 +172,13 @@ defmodule MapSet do
     %{map_set | map: Map.drop(map1, Map.keys(map2))}
   end
 
-  defp filter_not_in([], _map2, acc), do: :maps.from_list(acc)
+  defp filter_not_in(:none, _map2, acc), do: :maps.from_list(acc)
 
-  defp filter_not_in([key | rest], map2, acc) do
+  defp filter_not_in({key, _val, iter}, map2, acc) do
     if :erlang.is_map_key(key, map2) do
-      filter_not_in(rest, map2, acc)
+      filter_not_in(:maps.next(iter), map2, acc)
     else
-      filter_not_in(rest, map2, [{key, @dummy_value} | acc])
+      filter_not_in(:maps.next(iter), map2, [{key, @dummy_value} | acc])
     end
   end
 
@@ -197,14 +198,18 @@ defmodule MapSet do
     {map1, map2} = order_by_size(map1, map2)
 
     map1
-    |> Map.keys()
+    |> :maps.iterator()
+    |> :maps.next()
     |> none_in?(map2)
   end
 
   @compile {:inline, [none_in?: 2]}
 
-  defp none_in?([], _), do: true
-  defp none_in?([key | rest], map2), do: not :erlang.is_map_key(key, map2) and none_in?(rest, map2)
+  defp none_in?(:none, _), do: true
+
+  defp none_in?({key, _val, iter}, map2) do
+    not :erlang.is_map_key(key, map2) and none_in?(:maps.next(iter), map2)
+  end
 
   @doc """
   Checks if two sets are equal.
@@ -227,7 +232,7 @@ defmodule MapSet do
   # Elixir v1.5 change the map representation, so on
   # version mismatch we need to compare the keys directly.
   def equal?(%MapSet{map: map1}, %MapSet{map: map2}) do
-    map_size(map1) == map_size(map2) and map_subset?(Map.keys(map1), map2)
+    map_size(map1) == map_size(map2) and all_in?(map1, map2)
   end
 
   @doc """
@@ -309,19 +314,22 @@ defmodule MapSet do
   """
   @spec subset?(t, t) :: boolean
   def subset?(%MapSet{map: map1}, %MapSet{map: map2}) do
-    if map_size(map1) <= map_size(map2) do
-      map1
-      |> Map.keys()
-      |> map_subset?(map2)
-    else
-      false
-    end
+    map_size(map1) <= map_size(map2) and all_in?(map1, map2)
   end
 
-  defp map_subset?([], _), do: true
+  @compile {:inline, [all_in?: 2]}
 
-  defp map_subset?([key | rest], map2) do
-    :erlang.is_map_key(key, map2) and map_subset?(rest, map2)
+  defp all_in?(:none, _), do: true
+
+  defp all_in?({key, _val, iter}, map2) do
+    :erlang.is_map_key(key, map2) and all_in?(:maps.next(iter), map2)
+  end
+
+  defp all_in?(map1, map2) when is_map(map1) and is_map(map2) do
+    map1
+    |> :maps.iterator()
+    |> :maps.next()
+    |> all_in?(map2)
   end
 
   @doc """