@josevalim 2019-10-04
The difference for large sets happen because of the upfront cost in getting all keys. For a fair comparison, you also need to measure the cost of these functions when both sets are equal, that’s our worst case scenario. Can you please do so?

In any case, I believe these changes are beneficial. I have recently submitted a patch to Erlang mapsets with similar changes, and we should probably do the same here. However, we will need to wait until we drop Erlang 20. We are theoretically waiting for v1.10 to be out but we can bring it earlier depending on @hauleth’s work on the logger integration.