@sorentwo 2019-10-04
Totally understood that we need to drop Erlang 20 first. I thought (hoped?) it was time for that based on the logger integration work and a few other issues recently.

As requested, I ran a more detailed benchmark with worst case inputs of 10 through 1 million

```
Operating System: macOS
CPU Information: Intel(R) Core(TM) i5-7360U CPU @ 2.30GHz
Number of Available Cores: 4
Available memory: 8 GB
Elixir 1.9.0
Erlang 22.0

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
parallel: 1
inputs: 10, 100, 1000, 10000, 100000, 1000000
Estimated total run time: 1.40 min

Benchmarking 1.9.0 with input 10...
Benchmarking 1.9.0 with input 100...
Benchmarking 1.9.0 with input 1000...
Benchmarking 1.9.0 with input 10000...
Benchmarking 1.9.0 with input 100000...
Benchmarking 1.9.0 with input 1000000...
Benchmarking Custom with input 10...
Benchmarking Custom with input 100...
Benchmarking Custom with input 1000...
Benchmarking Custom with input 10000...
Benchmarking Custom with input 100000...
Benchmarking Custom with input 1000000...

##### With input 10 #####
Name             ips        average  deviation         median         99th %
1.9.0         1.39 M        0.72 μs ±31532.56%           0 μs           1 μs
Custom        0.95 M        1.05 μs ±25070.78%           0 μs           1 μs

Comparison:
1.9.0         1.39 M
Custom        0.95 M - 1.46x slower +0.33 μs

##### With input 100 #####
Name             ips        average  deviation         median         99th %
Custom      364.37 K        2.74 μs ±15511.18%           0 μs           1 μs
1.9.0       266.38 K        3.75 μs ±12572.16%           0 μs           2 μs

Comparison:
Custom      364.37 K
1.9.0       266.38 K - 1.37x slower +1.01 μs

##### With input 1000 #####
Name             ips        average  deviation         median         99th %
Custom      357.60 K        2.80 μs ±15320.90%           0 μs           1 μs
1.9.0        26.68 K       37.48 μs  ±3339.23%           6 μs          19 μs

Comparison:
Custom      357.60 K
1.9.0        26.68 K - 13.40x slower +34.68 μs

##### With input 10000 #####
Name             ips        average  deviation         median         99th %
Custom      355.01 K        2.82 μs ±15286.94%           0 μs           1 μs
1.9.0         2.49 K      402.25 μs   ±944.93%         100 μs         184 μs

Comparison:
Custom      355.01 K
1.9.0         2.49 K - 142.80x slower +399.44 μs

##### With input 100000 #####
Name             ips        average  deviation         median         99th %
Custom      398.34 K     0.00251 ms ±16339.24%           0 ms     0.00100 ms
1.9.0         0.20 K        4.96 ms   ±282.56%        1.20 ms       60.25 ms

Comparison:
Custom      398.34 K
1.9.0         0.20 K - 1975.07x slower +4.96 ms

##### With input 1000000 #####
Name             ips        average  deviation         median         99th %
Custom      351.08 K     0.00285 ms ±20803.20%           0 ms     0.00100 ms
1.9.0       0.0221 K       45.33 ms   ±122.39%       15.71 ms      203.02 ms

Comparison:
Custom      351.08 K
1.9.0       0.0221 K - 15912.62x slower +45.32 ms
```

Here is the benchmark itself:

```elixir
defmodule Custom do
  def disjoint?(%MapSet{map: map1}, %MapSet{map: map2}) do
    {map1, map2} = order_by_size(map1, map2)

    map1
    |> :maps.iterator()
    |> :maps.next()
    |> none_in?(map2)
  end

  @compile {:inline, [none_in?: 2]}

  defp none_in?(:none, _), do: true

  defp none_in?({key, _val, iter}, map2) do
    not :erlang.is_map_key(key, map2) and none_in?(:maps.next(iter), map2)
  end

  @compile {:inline, [order_by_size: 2]}
  defp order_by_size(map1, map2) when map_size(map1) > map_size(map2), do: {map2, map1}
  defp order_by_size(map1, map2), do: {map1, map2}
end

make_set = fn n ->
  for index <- 0..n, into: MapSet.new() do
    :md5 |> :crypto.hash(to_string(index)) |> Base.encode64()
  end
end

Benchee.run(%{
  "1.9.0" => fn set -> MapSet.disjoint?(set, set) end,
  "Custom" => fn set -> Custom.disjoint?(set, set) end
}, inputs: %{
  "10" => make_set.(10),
  "100" => make_set.(100),
  "1000" => make_set.(1000),
  "10000" => make_set.(10000),
  "100000" => make_set.(100000),
  "1000000" => make_set.(1000000)
})
```