@josevalim 2019-10-04
Oh wow, very interesting. So even for same inputs it is much faster. I guess it makes sense: we don’t have to allocate such a big list. So your initial benchmark was on worst case scenario all along then?

I wonder if we should also change the Enumerable implementation to use iterators. But that’s definitely a separate discussion.