@josevalim 2019-10-04
And thank you for running further benchmarks!