@michalmuskala 2019-10-04
Iterators make sense here since we don't need to do the unpack-repack dance with tuples we'd need to do in `Enumerable`.